tools/validation: Move manual validation code into ./validation
